Added rest of day 7 solution

+++ Solutions/1.cs
@@ -8,7 +8,7 @@ public class Day1(IFileLoader loader, IOptions<SolutionOptions> options) : ISolu
 {
     public int Day => 1;
 
-    public async Task<int> Solve()
+    public async Task<long> Solve()
     {
         var lines = await loader.Load<int>(Day, options.Value.SolutionType, options.Value.RunType);
         var listOne = lines.Select(parts => parts.First()).ToList();
