+++ Solutions/2.cs
@@ -8,7 +8,7 @@ public class Day2(IFileLoader loader, IOptions<SolutionOptions> options) : ISolu
 {
     public int Day => 2;
 
-    public async Task<int> Solve()
+    public async Task<long> Solve()
     {
         var reports = await loader.Load<int>(Day, options.Value.SolutionType, options.Value.RunType);
         var safeCount = reports.Count(x => options.Value.SolutionType switch
